+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - perf: add updateShape function for donut node, closes: #4091;
 - chore: refactor the implementation of donut node;
+- feat: legend supports diamond, ellipse, triangle, star shape, closes: #3876;
 
 ### 4.8.1